Make responsive index
List of changes
Make the header to be centered (moving menú ítems to correct position using height: 100%)
Make the logo a Little bit to the top to center its text (top: -11px)
Move the hamburger menu down by top:1.65% so it doesn't cover the header and the close button
Make the phone number button at the header to use flex-center instead of flex-end,  so it centers accordignly to the header height
Add a media query for tab menu where it changes to flex-wrap, overriding the overflow and make the tabs to wrap accordingly to screen size
Add a z-index to w-nav-button (hamburger) to z-index: 2 to make it clickable over the menu (it was clickable before but cursor pointer wasn't working)
Add a new class "text-cta" to edit text in that section without affecting the styles of paragraph.light
Remove display none from bold-text-9
Add flex auto and width auto to paragraph.light.text-cta in screens of 991px
Remove inset from rectangle-cta in screens of 991px
Add media query in screens of 850px for text-cta to center
Move title-wrapper higher in hierarchy to make the content adjust from column to row using flex-direction: row in screens of 767 px
Make paragraph about-us to have a width auto in screens of 767px
Make numbers-frame to have flex-direction column and min width of 3% to change the presentation to the right in screens of 767px
Make feature-wrap to have flex-direction row, gap 20px and align-items to center in screens of 767px
Change HTML <p> to multiple <p> and remove <br> to improve SEO (it also makes easy to apply styles)
Make paragraph inside feature-wrap to have margin left and padding left of 0.
Change div "numbers" to have <p> inside, this to make easy the position to center horizontally and improves SEO
Make numbers div to have a min-width of 50px, display flex, align-items center, justify-content center, top 10% (it is to make it centered Vertically) and position relative.
Make paragraph narrow to have top 111px only to the hero section
Increase rectangle-contact width to 160px and move more to the left to adjust accordingly to size in 991px
Add media query in 600px to numbers frame, increase min-width to 40%, feature-wrap reduce margin bottom and gap to 10px and make paragraph about-us to have padding right of 10px
Remove bold-text styles from media query 479px
Remove container home-hero styles from media query 479px
Make logo to have width 100% in screens 479px
Make numbers and paragraph from feature-wrap to decrease font size in 479px
Make navbar button to have width auto and max-width of 50% in screens of 767px